fix(tests): fix broken tests
