+++ .github/workflows/tests.yml
@@ -83,79 +83,77 @@ jobs:
           npm ci --no-optional
           npm run lint:markdown
 
-#  tests:
-#    name: Meteor ${{ matrix.meteor }} tests
-#    runs-on: ubuntu-latest
-#    steps:
-#      - name: checkout
-#        uses: actions/checkout@v3
-#
-#      - name: Install Meteor
-#        id: cache-meteor-install
-#        uses: actions/cache@v3
-#        with:
-#          path: ~/.meteor
-#          key: v1-meteor-${{ hashFiles('.meteor/versions') }}
-#          restore-keys: |
-#            v1-meteor-
-#      - name: Cache NPM dependencies
-#        id: cache-meteor-npm
-#        uses: actions/cache@v3
-#        with:
-#          path: ~/.npm
-#          key: ${{ runner.os }}-node-14-${{ hashFiles('**/package-lock.json') }}
-#          restore-keys: |
-#            ${{ runner.os }}-node-14-
-#      - run: |
-#          cd src
-#          npm ci
-#
-#      - name: Cache Meteor build
-#        id: cache-meteor-build
-#        uses: actions/cache@v3
-#        with:
-#          path: |
-#            .meteor/local/resolver-result-cache.json
-#            .meteor/local/plugin-cache
-#            .meteor/local/isopacks
-#            .meteor/local/bundler-cache/scanner
-#          key: v1-meteor_build_cache-${{ github.ref }}-${{ github.sha }}
-#          restore-keys: |
-#            v1-meteor_build_cache-
-#      - name: Setup meteor
-#        uses: meteorengineer/setup-meteor@v1
-#        with:
-#          meteor-release: '2.7.3'
-#
-#        # uses -o for running once -c for coverage
-#      - name: Run Tests
-#        run: |
-#          cd src
-#          sh ./scripts/test.sh -c -o
-#
-#      - name: Upload coverage
-#        uses: actions/upload-artifact@v2
-#        with:
-#          name: coverage-folder
-#          path: src/.coverage/
-#
-#  coverage:
-#    name: Coverage report
-#    runs-on: ubuntu-latest
-#    needs: [ tests ]
-#    steps:
-#      - name: Checkout
-#        uses: actions/checkout@v3
-#
-#      - name: Download coverage
-#        uses: actions/download-artifact@v2
-#        with:
-#          name: coverage-folder
-#          path: src/.coverage/
-#
-#
-#      - name: Coverage Report
-#        uses: VeryGoodOpenSource/very_good_coverage@v1.1.1
-#        with:
-#          path: ".coverage/lcov.info"
-#          min_coverage: 65 # TODO increase to 95!
+  tests:
+    name: Meteor ${{ matrix.meteor }} tests
+    runs-on: ubuntu-latest
+    steps:
+      - name: checkout
+        uses: actions/checkout@v3
+
+      - name: Install Meteor
+        id: cache-meteor-install
+        uses: actions/cache@v3
+        with:
+          path: ~/.meteor
+          key: v1-meteor-${{ hashFiles('.meteor/versions') }}
+          restore-keys: |
+            v1-meteor-
+      - name: Cache NPM dependencies
+        id: cache-meteor-npm
+        uses: actions/cache@v3
+        with:
+          path: ~/.npm
+          key: ${{ runner.os }}-node-14-${{ hashFiles('**/package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-node-14-
+
+      - name: Cache Meteor build
+        id: cache-meteor-build
+        uses: actions/cache@v3
+        with:
+          path: |
+            .meteor/local/resolver-result-cache.json
+            .meteor/local/plugin-cache
+            .meteor/local/isopacks
+            .meteor/local/bundler-cache/scanner
+          key: v1-meteor_build_cache-${{ github.ref }}-${{ github.sha }}
+          restore-keys: |
+            v1-meteor_build_cache-
+      - name: Setup meteor
+        uses: meteorengineer/setup-meteor@v1
+        with:
+          meteor-release: '2.7.3'
+
+        # uses -o for running once -c for coverage
+      - name: Run Tests
+        run: |
+          cd src
+          meteor npm ci --no-optional
+          sh ./scripts/test.sh -c -o -p
+
+      - name: Upload coverage
+        uses: actions/upload-artifact@v2
+        with:
+          name: coverage-folder
+          path: src/.coverage/
+
+  coverage:
+    name: Coverage report
+    runs-on: ubuntu-latest
+    needs: [ tests ]
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      - name: Download coverage
+        uses: actions/download-artifact@v2
+        with:
+          name: coverage-folder
+          path: src/.coverage/
+
+
+      - name: Coverage Report
+        uses: VeryGoodOpenSource/very_good_coverage@v1.1.1
+        with:
+          path: ".coverage/lcov.info"
+          min_coverage: 65 # TODO increase to 95!
