+++ src/scripts/test.sh
@@ -17,7 +17,8 @@ T_RUN_ONCE=""           # runs in watch mode
 T_VERBOSE=0             # no extra verbosity
 T_SERVER=1              # runs server tests
 T_CLIENT=1              # runs client tests
-
+T_PARALLEL=0            # runs tests in parallel
+T_LEGACY="--exclude-archs web.browser.legacy,web.cordova"             # exclude legacy archs
 # options:
 
 SCRIPT_USAGE="
@@ -29,22 +30,24 @@ Options:
   -c              Activate code-coverage reports
   -g <RegExp>     Filter tests by a given RegExp (uses Mocha-grep)
   -h              Show help
+  -l              Include legacy architecture (web.browser.legacy)
   -o              Runs the tests only once (default is watch-mode)
+  -p              Runs tests in parallel
   -v              Verbose mode with extra prints
 "
 
 
-while getopts "a:bcg:hov" opt; do
+while getopts "a:bcg:hlopv" opt; do
   case $opt in
     a)
       if [ "$OPTARG" = "client" ]
       then
-        T_CLIENT=1
-        T_SERVER=0
+        T_CLIENT="1"
+        T_SERVER="0"
       elif [ "$OPTARG" = "server" ]
       then
-        T_CLIENT=0
-        T_SERVER=1
+        T_CLIENT="0"
+        T_SERVER="1"
       else
         echo "Invalid parameter value for -a: $OPTARG"
         echo "$SCRIPT_USAGE"
@@ -58,13 +61,19 @@ while getopts "a:bcg:hov" opt; do
       T_FILTER=${OPTARG}
       ;;
     v)
-	  T_VERBOSE=1
+	  T_VERBOSE="1"
       ;;
     c)
-      T_COVERAGE=1
+      T_COVERAGE="1"
+      ;;
+    l)
+      T_LEGACY="--exclude-archs web.cordova"
       ;;
     o)
       T_RUN_ONCE="--once"
+      ;;
+    p)
+      T_PARALLEL="1"
       ;;
     h)
       echo "$SCRIPT_USAGE"
@@ -97,6 +106,7 @@ then
 	echo "=> grep pattern: [${T_FILTER}]"
 	echo "=> coverage: [${T_COVERAGE}]"
 	echo "=> Browser: [${T_BROWSER}]"
+	echo "=> Parallel: [${T_PARALLEL}]"
 	echo "=> Arch: [server: ${T_SERVER}, client: ${T_CLIENT}]"
 fi
 
@@ -107,6 +117,9 @@ METEOR_PACKAGE_DIRS=${T_PACKAGE_DIRS}  \
     TEST_SERVER=${T_SERVER} \
     TEST_CLIENT=${T_CLIENT} \
     MOCHA_GREP=${T_FILTER} \
+    TEST_PARALLEL=${T_PARALLEL} \
+    SERVER_TEST_REPORTER="spec" \
+    CLIENT_TEST_REPORTER="spec" \
     BABEL_ENV=COVERAGE \
     COVERAGE=${T_COVERAGE} \
     COVERAGE_OUT_HTML=1 \
@@ -115,6 +128,7 @@ METEOR_PACKAGE_DIRS=${T_PACKAGE_DIRS}  \
     COVERAGE_VERBOSE_MODE=${T_VERBOSE} \
     meteor test \
         ${T_RUN_ONCE} \
+        ${T_LEGACY} \
         --driver-package=meteortesting:mocha \
         --settings=tests/settings.json \
         --port=${PORT}
