+++ src/tests/client/main.js
@@ -10,23 +10,26 @@ import '../../imports/infrastructure/pipelines/tests'
 
 /// /////////////////////////////////////////////////////////////////////
 //
-// DEFAULTS
+// API
 //
 /// /////////////////////////////////////////////////////////////////////
-// we need to import the minimal defaults that are required for everything
-// else to work properly
-
-import '../../imports/startup/client/minimal/defaults'
-import '../../imports/utils/tests/archUtils.test'
-import '../../imports/contexts/beamer/tests'
-import '../../imports/contexts/system/accounts/users/tests/UserUtils.tests'
+import '../../imports/api/schema/tests'
 import '../../imports/api/accounts/registration/tests/PasswordConfig.tests'
 import '../../imports/api/accounts/registration/tests/registerUserSchema'
+
+/// /////////////////////////////////////////////////////////////////////
+//
+// STARTUP
+//
+/// /////////////////////////////////////////////////////////////////////
+import '../../imports/startup/client/minimal/defaults'
+import '../../imports/utils/tests/archUtils.test'
+
+import '../../imports/contexts/beamer/tests'
+import '../../imports/contexts/system/accounts/users/tests/UserUtils.tests'
 import '../../imports/contexts/classroom/invitations/tests/CodeInvitation.tests'
 import '../../imports/contexts/classroom/lessons/tests/LessonStates.tests'
-import '../../imports/contexts/classroom/lessons/tests/Lesson.tests'
-
-import '../../imports/api/schema/tests/Resolvers.tests'
+import '../../imports/contexts/classroom/lessons/tests/LessonHelpers.tests'
 
 // UI
 import '../../imports/ui/tests'
